+++ plugins/renderer/opengl/src/OpenGL/openglmeshloader.h
@@ -10,13 +10,13 @@ namespace Render {
 
         std::shared_ptr<OpenGLMeshData> loadImpl(ResourceType *res) override;
 
-        static OpenGLMeshData generate(size_t groupSize, Vertex *vertices, size_t vertexCount, unsigned int *indices = nullptr, size_t indexCount = 0);
-        static OpenGLMeshData generate(size_t groupSize, Vertex2 *vertices, size_t vertexCount, unsigned int *indices = nullptr, size_t indexCount = 0);
-        static OpenGLMeshData generate(size_t groupSize, Vertex3 *vertices, size_t vertexCount, unsigned int *indices = nullptr, size_t indexCount = 0);
+        static OpenGLMeshData generate(size_t groupSize, Vertex *vertices, size_t vertexCount, unsigned short *indices = nullptr, size_t indexCount = 0);
+        static OpenGLMeshData generate(size_t groupSize, Vertex2 *vertices, size_t vertexCount, unsigned short *indices = nullptr, size_t indexCount = 0);
+        static OpenGLMeshData generate(size_t groupSize, Vertex3 *vertices, size_t vertexCount, unsigned short *indices = nullptr, size_t indexCount = 0);
 
-        static void update(OpenGLMeshData &data, size_t groupSize, Vertex *vertices, size_t vertexCount, unsigned int *indices = nullptr, size_t indexCount = 0);
-        static void update(OpenGLMeshData &data, size_t groupSize, Vertex2 *vertices, size_t vertexCount, unsigned int *indices = nullptr, size_t indexCount = 0);
-        static void update(OpenGLMeshData &data, size_t groupSize, Vertex3 *vertices, size_t vertexCount, unsigned int *indices = nullptr, size_t indexCount = 0);
+        static void update(OpenGLMeshData &data, size_t groupSize, Vertex *vertices, size_t vertexCount, unsigned short *indices = nullptr, size_t indexCount = 0);
+        static void update(OpenGLMeshData &data, size_t groupSize, Vertex2 *vertices, size_t vertexCount, unsigned short *indices = nullptr, size_t indexCount = 0);
+        static void update(OpenGLMeshData &data, size_t groupSize, Vertex3 *vertices, size_t vertexCount, unsigned short *indices = nullptr, size_t indexCount = 0);
     };
 
 }
